+++ src/components/UserPortal/UserSidebarOrg/UserSidebarOrg.tsx
@@ -12,7 +12,7 @@ import type { TargetsType } from 'state/reducers/routesReducer';
 // import type { InterfaceQueryOrganizationsListObject } from 'utils/interfaces';
 // import AngleRightIcon from 'assets/svgs/angleRight.svg?react';
 import TalawaLogo from 'assets/svgs/talawa.svg?react';
-import styles from '../../../style/app.module.css';
+import styles from '../../../style/app-fixed.module.css';
 // import Avatar from 'components/Avatar/Avatar';
 import ProfileCard from 'components/ProfileCard/ProfileCard';
 import SignOut from './../../SignOut/SignOut';
