+++ src/screens/OrgPost/OrgPost.tsx
@@ -17,7 +17,7 @@ import { toast } from 'react-toastify';
 import convertToBase64 from 'utils/convertToBase64';
 import { errorHandler } from 'utils/errorHandler';
 import type { InterfaceQueryOrganizationPostListItem } from 'utils/interfaces';
-import styles from '../../style/app.module.css';
+import styles from '../../style/app-fixed.module.css';
 import SortingButton from '../../subComponents/SortingButton';
 import SearchBar from 'subComponents/SearchBar';
