+++ src/screens/OrganizationActionItems/ItemUpdateStatusModal.tsx
@@ -2,7 +2,7 @@ import React, { type FC, type FormEvent, useEffect, useState } from 'react';
 import { Modal, Button, Form } from 'react-bootstrap';
 import { useTranslation } from 'react-i18next';
 import { FormControl, TextField } from '@mui/material';
-import styles from '../../style/app.module.css';
+import styles from '../../style/app-fixed.module.css';
 import { useMutation } from '@apollo/client';
 import { UPDATE_ACTION_ITEM_MUTATION } from 'GraphQl/Mutations/ActionItemMutations';
 import { toast } from 'react-toastify';
